requirements.txt
django>=2.2.5
djangorestframework>=3.10.3
markdown>=3.1.1
gunicorn>=19.9.0

